+++ engine/execution/state/storehouse/sync/storehouse_test.go
@@ -1,4 +1,4 @@
-package cargo_test
+package sync_test
 
 import (
 	"math/rand"
@@ -6,13 +6,13 @@ import (
 	"testing"
 	"time"
 
-	"github.com/onflow/flow-go/engine/execution/state/cargo"
-	"github.com/onflow/flow-go/engine/execution/state/cargo/storage"
+	"github.com/onflow/flow-go/engine/execution/state/storehouse/storage"
+	"github.com/onflow/flow-go/engine/execution/state/storehouse/sync"
 	"github.com/onflow/flow-go/utils/unittest"
 	"github.com/stretchr/testify/require"
 )
 
-func TestCargo(t *testing.T) {
+func TestSynchronizer(t *testing.T) {
 
 	t.Run("happy path", func(t *testing.T) {
 
@@ -21,7 +21,7 @@ func TestCargo(t *testing.T) {
 
 		syncFreq := 50 * time.Millisecond
 		store := storage.NewInMemoryStorage(100, genesis, nil)
-		c, err := cargo.NewCargo(store, 100, genesis, syncFreq)
+		c, err := sync.NewSynchronizer(store, 100, genesis, syncFreq)
 
 		require.NoError(t, err)
